MOBPLAT-380 convert junit time field to float

+++ subprojects/test-runner/instrumentation-tests/src/main/kotlin/com/avito/instrumentation/internal/finalizer/verdict/LegacyTestStatisticsCounter.kt
@@ -9,8 +9,8 @@ internal class LegacyTestStatisticsCounter(
     val notReported: HasNotReportedTestsDeterminer.Result
 ) : TestStatisticsCounter {
 
-    override fun overallDurationSec(): Int {
-        return reportedTests.sumBy { it.lastAttemptDurationInSeconds }
+    override fun overallDurationSec(): Float {
+        return reportedTests.sumBy { it.lastAttemptDurationInSeconds }.toFloat()
     }
 
     override fun overallCount(): Int {
