+++ subprojects/test-runner/instrumentation-tests/src/main/kotlin/com/avito/instrumentation/internal/finalizer/verdict/TestStatisticsCounterImpl.kt
@@ -6,12 +6,12 @@ import com.avito.report.model.duration
 
 internal class TestStatisticsCounterImpl(private val verdict: Verdict) : TestStatisticsCounter {
 
-    override fun overallDurationSec(): Int =
+    override fun overallDurationSec(): Float =
         verdict.testResults
             .filterIsInstance<TestRuntimeData>()
             .map { it.duration }
             .sum()
-            .toInt()
+            .toFloat()
 
     override fun overallCount(): Int = verdict.testResults.size
